requirements.txt
mutagen==1.47.0
pillow==10.2.0
pydub==0.25.1
